requirements.txt
streamlit
scikit-learn
pandas
numpy
joblib
pdfplumber
pypdf
python-docx
transformers
torch
plotly
spacy
